11.09

+++ 1 sem/lab/lab_01/lab_01_homework/polovianiuk_1_11.py
@@ -1,4 +1,4 @@
-a, b, c, d, f = [int(x) for x in input().split()]
+a, b, c, d, f = [float(x) for x in input().split()]
 
 if 0 < a <= 100 and 0 < b <= 100 and 0 < c <= 100 and 0 < d <= 100 and 0 < f <= 100:
     s1 = (a + b + f) / 2
